* config/i386/mframe-{freebsd,generic,linux-gnu,openbsd,sysv4.2MP}.h, config/powerpc/mframe-{generic,linux-gnu}.h, config/sparc/mframe-{generic,sunos4.1.3}.h: Add casts for isdigit args.
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@4128 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/libobjc/config/i386/mframe-freebsd.h
@@ -30,7 +30,7 @@
   (TYPE) = objc_skip_typespec(type); \
   sprintf((DEST), "%.*s%d", (TYPE)-type, type, (CUM)); \
   if (*(TYPE) == '+') (TYPE)++; \
-  while (isdigit(*(TYPE))) \
+  while (isdigit((int) *(TYPE))) \
     { \
       (TYPE)++; \
     } \
