+++ swarm/libobjc/config/sparc/mframe-generic.h
@@ -65,7 +65,7 @@ struct sparc_args {
     { \
       (TYPE)++; \
     } \
-  while (isdigit(*(TYPE))) \
+  while (isdigit ((int) *(TYPE))) \
     { \
       (TYPE)++; \
     } \
